requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
Django==4.1.5
django-analytical==3.1.0
django-cleanup==6.0.0
django-phonenumber-field==7.0.2
docopt==0.6.2
et-xmlfile==1.1.0
gunicorn==20.1.0
huepy==1.2.1
openpyxl==3.0.10
pathlib==1.0.1
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2-binary==2.9.5
soupsieve==2.3.2.post1
sqlparse==0.4.3
